requirements.txt
streamlit
ollama
pypdf
faiss-cpu
numpy
